requirements.txt
django
django-allauth
django-rest-auth
djangorestframework
psycopg2-binary
pygments
markdown
coreapi
drf-nested-routers